Refactor: Move info card and table to design system
