+++ packages/extension/src/view/design-system/components/infoCard/index.tsx
