+++ packages/extension/src/view/design-system/components/infoCard/learnMoreDropdown.tsx
