+++ packages/extension/src/view/design-system/components/infoCard/tests/infoCard.test.tsx
