+++ packages/extension/src/view/design-system/components/table/tableHeader/headerResizer.tsx
