+++ packages/extension/src/view/devtools/components/antiCovertTracking/bounceTracking/index.tsx
@@ -21,7 +21,7 @@ import React from 'react';
 /**
  * Internal dependencies.
  */
-import InfoCard from '../../infoCard';
+import InfoCard from '../../../../design-system/components/infoCard';
 import { PSInfoKey } from '../../../../../utils/fetchPSInfo';
 
 const BounceTracking = () => {
