+++ packages/extension/src/view/devtools/components/privateAdvertising/attribution/index.tsx
@@ -21,7 +21,7 @@ import React from 'react';
 /**
  * Internal dependencies.
  */
-import InfoCard from '../../infoCard';
+import InfoCard from '../../../../design-system/components/infoCard';
 import { PSInfoKey } from '../../../../../utils/fetchPSInfo';
 
 const Attribution = () => {
